Update ClamAV component

+++ components/applications/patina_applications_clamav.sh
@@ -1,28 +1,28 @@
 #!/usr/bin/env bash
 
-##########
-# Notice #
-##########
+###########
+# LICENSE #
+###########
 
 # Patina: A 'patina', 'layer', or 'toolbox' for BASH under Linux.
 # Copyright (C) 2020 William Willis Whinn
 
-# This program is free software: you can redistribute it and/or modify
-# it under the terms of the GNU General Public License as published by
-# the Free Software Foundation, either version 3 of the License, or
-# (at your option) any later version.
+# This program is free software: you can redistribute it and/or modify it under the terms of the GNU
+# General Public License as published by the Free Software Foundation, either version 3 of the
+# License, or (at your option) any later version.
 
-# This program is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-# GNU General Public License for more details.
+# This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without
+# even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
+# General Public License for more details.
 
-# You should have received a copy of the GNU General Public License
-# along with this program.  If not, see <http://www.gnu.org/licenses/>.
+# You should have received a copy of the GNU General Public License along with this program. If not,
+# see <http://www.gnu.org/licenses/>.
 
 #############
-# Functions #
+# FUNCTIONS #
 #############
+
+# PATINA > FUNCTIONS > APPLICATIONS > CLAMAV/FRESHCLAM
 
 patina_clamav() {
   local patina_create_clamav_logfile
@@ -33,24 +33,32 @@ patina_clamav() {
 
   # Success: Display contents of help file.
   if [ "$1" = '--help' ] ; then
-    echo_wrap "Usage: p-clamscan [FILE/DIRECTORY] [OPTION]"
-    echo_wrap "Dependencies: 'clamscan' command from package 'clamav'."
-    echo_wrap "Warning: Command(s) may require 'sudo' password."
-    echo_wrap "Perform a recursive virus scan of a given location and record results."
+    echo -e "Usage: p-clamscan [FILE/DIRECTORY] [OPTION]"
+    echo -e "Dependencies: 'clamscan' command from package 'clamav'."
+    echo -e "Dependencies: 'freshclam' command from package 'clamav'/'clamav-update'."
+    echo -e "Warning: Command(s) may require 'sudo' password."
+    echo -e "Perform a recursive virus scan of a given location and record results."
     echo
-    echo_wrap "  --repair\tPurge and replace current virus database."
-    echo_wrap "  --help\tDisplay this help and exit."
+    echo -e "  --repair\tPurge and replace current virus database."
+    echo -e "  --help\tDisplay this help and exit."
     echo
     return 0
 
   # Failure: Command 'clamscan' is not available.
   elif ( ! command -v 'clamscan' > /dev/null 2>&1 ) ; then
     patina_raise_exception 'PE0006'
-    return 1
+    return 127
+
+  # Failure: Command 'freshclam' is not available.
+  elif ( ! command -v 'freshclam' > /dev/null 2>&1 ) ; then
+    patina_raise_exception 'PE0006'
+    return 127
 
   # Failure: Patina has not been given an argument.
   elif [ "$#" -eq 0 ] ; then
     patina_raise_exception 'PE0001'
+    echo
+    patina_clamav --help
     return 1
 
   # Failure: Patina has been given too many arguments.
@@ -59,14 +67,30 @@ patina_clamav() {
     return 1
 
   # Success: Repair Freshclam update mechanism.
-  # Warning: Uses 'sudo' to delete system files.
+  # Warning: Uses 'sudo' to delete system files and execute Freshclam update program.
   elif [ "$1" = '--repair' ] ; then
-    sudo rm --force \
-      "$clamav_path/bytecode.cvd" \
-      "$clamav_path/daily.cld" \
-      "$clamav_path/main.cvd" \
-      "$clamav_path/mirrors.dat"
-    return 0
+    printf "${YELLOW}This command requires root permissions. Continue [Y/N]?${COLOR_RESET} "
+    read -n1 -r answer
+    echo
+
+    case "$answer" in
+      'Y'|'y')
+        sudo rm --force \
+          "$clamav_path/bytecode.cvd" \
+          "$clamav_path/daily.cld" \
+          "$clamav_path/main.cvd" \
+          "$clamav_path/mirrors.dat" && \
+        sudo freshclam
+        return 0
+        ;;
+      'N'|'n')
+        return 0
+        ;;
+      *)
+        patina_raise_exception 'PE0003'
+        return 1
+        ;;
+    esac
 
   # Failure: Scan target does not exist.
   elif [[ ! -e "$1" ]] ; then
@@ -76,8 +100,9 @@ patina_clamav() {
   # Success: Guide user in performing virus scan.
   elif [ "$#" -ne 0 ] && [[ -e "$1" ]] ; then
     echo
-    printf "Do you wish to record a log file in your home directory [Y/N]? "
+    printf "${YELLOW}Do you wish to record a log file in your Home directory [Y/N]?${COLOR_RESET} "
     read -n1 -r answer
+    echo
 
     case "$answer" in
       'Y'|'y')
@@ -93,7 +118,7 @@ patina_clamav() {
         ;;
     esac
 
-    echo_wrap "\\n\\nPreparing 'clamav' virus scan, please wait...\\n"
+    echo -e "\\n\\nPreparing 'clamav' virus scan, please wait...\\n"
 
     case "$patina_create_clamav_logfile" in
       true)
@@ -120,8 +145,10 @@ patina_clamav() {
 }
 
 ###########
-# Aliases #
+# ALIASES #
 ###########
+
+# PATINA > FUNCTIONS > APPLICATIONS > CLAMAV/FRESHCLAM
 
 alias 'p-clamscan'='patina_clamav'
